+++ package.json
@@ -2694,6 +2694,7 @@
         "@blueprintjs/select": "3.10.0",
         "@jupyterlab/coreutils": "^3.1.0",
         "@jupyterlab/services": "^4.2.0",
+        "@koa/cors": "^3.0.0",
         "@msrvida/python-program-analysis": "^0.4.1",
         "ansi-regex": "^4.1.0",
         "arch": "^2.1.0",
@@ -2709,6 +2710,7 @@
         "iconv-lite": "^0.4.21",
         "inversify": "^4.11.1",
         "jsonc-parser": "^2.0.3",
+        "koa": "^2.11.0",
         "less-plugin-inline-urls": "^1.2.0",
         "line-by-line": "^0.1.6",
         "lodash": "^4.17.15",
@@ -2792,6 +2794,8 @@
         "@types/iconv-lite": "^0.0.1",
         "@types/jquery": "^1.10.35",
         "@types/jsdom": "^11.12.0",
+        "@types/koa": "^2.11.0",
+        "@types/koa__cors": "^2.2.3",
         "@types/loader-utils": "^1.1.3",
         "@types/lodash": "^4.14.104",
         "@types/md5": "^2.1.32",
